[DOC-769] [feature] - Datasource declaration using groovy builders(and magic) [DOC-709] [feature] - project.context renamed to project.properties
