+++ backend/src/test/resources/groovy/ExampleEnv.genesis
@@ -16,11 +16,6 @@ template {
     version("0.1")
     createWorkflow("create")
     destroyWorkflow("destroy")
-    dataSources {
-        dataSource("staticList").name("ds1").config([list:["value1",  "value2"]])
-        dataSource("staticList").name("ds2").config([list:["value3"]])
-        dataSource("dependentList").name("ds3").config([list: [1,3,4]])
-    }
 
 
     workflow("create") {
@@ -31,12 +26,6 @@ template {
             variable("test").description("test")
             variable("optional").description("Optional variable").optional(defaultValue = 1)
             variable("optionalNoValue").description("Optional variable without default value").optional()
-            variable("list").as(Integer).description("Variable that has one value from a list").oneOf {["1":1, "2":2, "3":3]}.optional()
-            variable("listDS1").as(String).description("Variable that has one value from DS").dataSource("ds1").optional()
-            variable("listDS12").as(String).description("Variable that has one value from DS").oneOf {ds1 + ds2}.optional()
-            variable("dependent").as(Character).description("Dependent variable").dependsOn("nodesCount").dataSource("ds3").optional()
-            variable("doubleDep").as(String).description("Double dependent variable").dependsOn("nodesCount").dependsOn("dependent").dataSource("ds3").optional()
-            variable("triple").as(String).description("Double dependent variable").dependsOn("list", "nodesCount", "dependent").dataSource("ds3").optional()
         }
 
         steps {
